requirements.txt
streamlit
ultralytics
pillow
pandas
numpy
cv2 
psycpog2